// src/types/resources/Appointment.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TUnsignedInt } from '../simpleTypes/UnsignedInt';
import { TInstant } from '../simpleTypes/Instant';
import { TInt } from '../simpleTypes/Int';
import { TDateTime } from '../simpleTypes/DateTime';
import { TAppointmentParticipant } from '../partials/AppointmentParticipant';
import { TPeriod } from '../partials/Period';

export type TAppointment = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  status: string;
  cancelationReason?: TCodeableConcept;
  serviceCategory?: TCodeableConcept[];
  serviceType?: TCodeableConcept[];
  specialty?: TCodeableConcept[];
  appointmentType?: TCodeableConcept;
  reasonCode?: TCodeableConcept[];
  reasonReference?: TReference[];
  priority?: TUnsignedInt;
  description?: string;
  supportingInformation?: TReference[];
  start?: TInstant;
  end?: TInstant;
  minutesDuration?: TInt;
  slot?: TReference[];
  created?: TDateTime;
  comment?: string;
  patientInstruction?: string;
  basedOn?: TReference[];
  participant: TAppointmentParticipant[];
  requestedPeriod?: TPeriod[];
};